% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{c_integer}
\alias{c_integer}
\title{Concatenates two integer vectors.}
\usage{
c_integer(x, y)
}
\arguments{
\item{x}{Integer vector.}

\item{y}{Integer vector.}
}
\description{
Concatenates two integer vectors.
}
\keyword{internal}
