% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_seed_sets.R
\name{check_seed_sets}
\alias{check_seed_sets}
\title{Verify that user supplied seed sets are valid.}
\usage{
check_seed_sets(seeds, G_stats)
}
\arguments{
\item{seeds}{List of seed sets. Each seed set should be a vector of integer vertex IDs of G.}

\item{G_stats}{List of graph statistics returned from \code{eval_G}.}
}
\value{
Possibly, seeds.
}
\description{
Verify that user supplied seed sets are valid.
}
\keyword{internal}
