% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_burn.R
\name{cluster_burn}
\alias{cluster_burn}
\title{Greedy method to extract statistically significant communities from a heterogeneous graph where nodes are differentiated by type and edges are distinguished from their incident node types.}
\usage{
cluster_burn(
  G,
  node_type,
  alpha,
  adj_method = "fdr",
  learning_rate = 1,
  decay_rate = 0.6,
  max_iter = igraph::vcount(G),
  loco_samples = FALSE
)
}
\arguments{
\item{G}{igraph representation of graph.}

\item{node_type}{Name of vertex attribute containing the node types.}

\item{alpha}{Numeric between 0 and 1 representing desired type I error rate.}

\item{nsamples}{Integer between 1 and the number of nodes in \code{G} representing the number of samples to be taken.}
}
\value{
A list with the extracted community structure.
}
\description{
Greedy method to extract statistically significant communities from a heterogeneous graph where nodes are differentiated by type and edges are distinguished from their incident node types.
}
