% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_ZCMod.R
\name{cluster_ZCMod}
\alias{cluster_ZCMod}
\title{Modularity maximization method to find dense, heterogeneous communities.}
\usage{
cluster_ZCMod(G, node_type, iter = 10)
}
\arguments{
\item{G}{igraph representation of graph.}

\item{node_type}{Name of vertex attribute containing the node types.}

\item{iter}{Number of iterations considered. The iteration with the highest modularity is chosen.}
}
\value{
A list with the communities and modularity measure.
}
\description{
Modularity maximization method to find dense, heterogeneous communities.
}
