% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_conductance.R
\name{compute_conductance}
\alias{compute_conductance}
\title{Compute conductance of a cut.}
\usage{
compute_conductance(B0, G_stats)
}
\arguments{
\item{B0}{Vector of vertices representing the candidate community.}

\item{G_stats}{List of graph statistics returned from \code{eval_G}.}
}
\value{
A double between 0 and 1.
}
\description{
Compute conductance of a cut.
}
\keyword{internal}
