% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_pvals_binom.R
\name{compute_pvals_binom}
\alias{compute_pvals_binom}
\title{Composite Binomial p-value for each node.}
\usage{
compute_pvals_binom(B0, G_stats)
}
\arguments{
\item{B0}{Vector of vertices representing the candidate community.}

\item{G_stats}{List of graph statistics returned from \code{eval_G}.}
}
\value{
A list with K elements. Each element represents the probability of connection to the candidate community in the unipartite and bipartite subgraphs.
}
\description{
Computes the binomial composite p-value for each node. This p-value
represents the probability of attaining at least as many connections to B in each
subgraph jointly than what is observed. Small values are indications that a given
node is more connected to the candidate community through provided node types than
what would be expected at random.
}
\keyword{internal}
