% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_zcmodularity_matrix.R
\name{compute_zcmodularity_matrix}
\alias{compute_zcmodularity_matrix}
\title{Computes the modularity matrix for an arbitrary heterogeneous graph.}
\usage{
compute_zcmodularity_matrix(G, node_type)
}
\arguments{
\item{G}{igraph representation of graph.}

\item{node_type}{Name of vertex attribute containing the node types.}
}
\value{
A matrix.
}
\description{
Computes the modularity matrix for an arbitrary heterogeneous graph.
}
