% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_affiliation.R
\name{convert_affiliation}
\alias{convert_affiliation}
\title{Convert node affiliation.}
\usage{
convert_affiliation(G, affiliation, method = c("c2m", "m2c"))
}
\arguments{
\item{G}{igraph representation of graph.}

\item{affiliation}{List of communities or membership list.}

\item{method}{String \code{c2m} or \code{m2c}.}
}
\value{
The converted community structure.
}
\description{
Converts a list of communities to a membership list (if overlapping) or vector (if nonoverlapping) using \code{c2m}. Converts a membership list or vector into a list of communities using \code{m2c}.
}
