% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{count_stubs_for_B0_cpp}
\alias{count_stubs_for_B0_cpp}
\title{Count stubs for p-value computation using theorem.}
\usage{
count_stubs_for_B0_cpp(B0, G_stats)
}
\arguments{
\item{B0}{Vector of vertices representing the candidate community.}

\item{G_stats}{List of graph statistics returned from \code{eval_G}.}
}
\description{
Count stubs for p-value computation using theorem.
}
\details{
Counts the number of colored stubs for each color emanating from each
node which connects to a provided set of nodes in the observed graph.
}
\keyword{internal}
