% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_stubs_on_v.R
\name{count_stubs_on_v}
\alias{count_stubs_on_v}
\title{Count stubs on each vertex by each subgraph.}
\usage{
count_stubs_on_v(n, K, nodes, adj_nodes, type_map)
}
\arguments{
\item{n}{Integer for number of nodes}

\item{K}{Integer for number of node types.}

\item{nodes}{Data frame with node column and type column providing the node type
for each node.}

\item{adj_nodes}{Adjacency list, generated with \code{igraph::adjacent_vertices}.}

\item{type_map}{Type map, generated with \code{create_type_map}.}
}
\value{
An integer matrix of size n by K where n is the number of nodes and K
is the number of node types.
}
\description{
A vertex of a given type is connected to other vertices through
stubs. The color of the stub indicates the color of the adjacent nodes. That is,
stubs can be partitioned based on color. The first column of the resulting
matrix provides the number of same type connections. All other columns provide
the number of between type connections whether the order is dictated by the
node type and the type map. The ith row provides the ith node's stub sequence.
}
\keyword{internal}
