% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_type_map.R
\name{create_type_map}
\alias{create_type_map}
\title{Creates a data frame outlining all possible within-type and between-type connections, labeling unique combinations.}
\usage{
create_type_map(G, node_type)
}
\arguments{
\item{G}{igraph representation of graph.}

\item{node_type}{name of vertex attribute containing the node types.}
}
\value{
A dataframe with unique connection combinations.
}
\description{
Creates a data frame outlining all possible within-type and between-type connections, labeling unique combinations.
}
