% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_community.R
\name{eval_community}
\alias{eval_community}
\title{Evaluate community.}
\usage{
eval_community(G, vids)
}
\arguments{
\item{G}{igraph representation of graph.}

\item{vids}{Vector of vertex ids representing a single community.}
}
\value{
Data frame.
}
\description{
Evaluates a community based on common metrics for connectedness.
Supports computation of edge densities within, between, and beyond the community as well as computation of local modularity.
}
