% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_G.R
\name{eval_G}
\alias{eval_G}
\title{Evaluate degree characteristics of G and compute other needed graphical statistics
and summaries.}
\usage{
eval_G(G, node_type)
}
\arguments{
\item{G}{igraph representation of graph.}

\item{node_type}{Name of vertex attribute containing the node types.}
}
\value{
A list with needed graph summaries, including: (1) the node set, (2)
edge set, (3) cardinality of the vertex set, (4) cardinality of the edge set,
(5) number of node types, (6) data frame with mapping between node and node type,
(7) adjacency list, and (8) the type map
}
\description{
Evaluate degree characteristics of G and compute other needed graphical statistics
and summaries.
}
\keyword{internal}
