% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_jaccard.R
\name{eval_jaccard}
\alias{eval_jaccard}
\title{Summarize and return grid of jaccard values.}
\usage{
eval_jaccard(truth, discovery)
}
\arguments{
\item{truth}{A named list of vectors representing true communities.}

\item{discovery}{A list of vectors representing discovered communities.}
}
\value{
A list with the best mappings and all possible mappings between truth
and discovery. See details.
}
\description{
Summarize and return grid of jaccard values.
}
\details{
Best is a data frame providing the discovered community which most
closely resembles each of the true communities and the average jaccard across
each mapping (truth to discovery and discovery to truth). The average is bounded
between 0 and 1. The mapped averages are bounded between 0 and 0.5.
The grid provides the jaccard for all possible truth and discovery pairings.
}
