% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_loco_egonets.R
\name{eval_loco_egonets}
\alias{eval_loco_egonets}
\title{Find nodes whose neighborhood is a locally optimal seed set.}
\usage{
eval_loco_egonets(G_stats)
}
\arguments{
\item{G_stats}{List of graph statistics returned from \code{eval_G}.}
}
\value{
Vector of nodes.
}
\description{
Find nodes whose neighborhood is a locally optimal seed set.
}
\keyword{internal}
