% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_ECoHeN.R
\name{extract_ECoHeN}
\alias{extract_ECoHeN}
\title{Extract communities from heterogeneous networks (ECoHeN) with user supplied
seed sets.}
\usage{
extract_ECoHeN(
  G,
  seeds,
  node_type,
  alpha,
  learning_rate = 1,
  decay_rate = 0.99,
  adj_method = "fdr",
  max_iter = igraph::vcount(G)
)
}
\arguments{
\item{G}{igraph representation of graph or results of \code{eval_G}.}

\item{seeds}{List of seed sets. Each seed set should be a vector of integer vertex IDs of G.}

\item{node_type}{Name of vertex attribute containing the node types.}

\item{alpha}{Numeric between 0 and 1 representing desired significance level.}

\item{learning_rate}{Numeric between 0 and 1 dictating the maximal allowance on the first iteration.}

\item{decay_rate}{Numeric between 0 and 1 dictating the maximal allowance after the first iteration.}

\item{adj_method}{Character scalar for the desired multiple correction method. See \code{p.adjust.methods} for options.}

\item{max_iter}{Integer scalar for the maximum number of iterations before halting the extraction procedure.}
}
\value{
A list with the extracted community structure.
}
\description{
Extract communities from heterogeneous networks (ECoHeN) with user supplied
seed sets.
}
\details{
Method to extract statistically significant communities from a heterogeneous
graph where nodes are differentiated by type and edges are distinguished from
their incident node types.
}
