% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{index_Cpp2R}
\alias{index_Cpp2R}
\title{Converts C++ indices to R indices by adding one.}
\usage{
index_Cpp2R(x)
}
\arguments{
\item{x}{Integer vector.}
}
\description{
Converts C++ indices to R indices by adding one.
}
\keyword{internal}
