% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{index_R2Cpp}
\alias{index_R2Cpp}
\title{Converts R indices to C++ by subtracting one.}
\usage{
index_R2Cpp(x)
}
\arguments{
\item{x}{Integer vector.}
}
\description{
Converts R indices to C++ by subtracting one.
}
\keyword{internal}
