% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{is_cycle}
\alias{is_cycle}
\title{Checks if the extraction procedure reached a cycle.}
\usage{
is_cycle(x, y, z)
}
\arguments{
\item{x}{An integer vector of nodes.}

\item{y}{An integer vector of nodes.}

\item{z}{An integer vector of nodes.}
}
\description{
Checks if the extraction procedure reached a cycle.
}
\keyword{internal}
