% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{is_empty}
\alias{is_empty}
\title{Checks if the extraction procedure reached empty.}
\usage{
is_empty(x)
}
\arguments{
\item{x}{An integer vector of nodes.}
}
\description{
Checks if the extraction procedure reached empty.
}
\keyword{internal}
