% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{is_stable}
\alias{is_stable}
\title{Checks if the extraction procedure converged.}
\usage{
is_stable(x, y, z)
}
\arguments{
\item{x}{An integer vector of nodes.}

\item{y}{An integer vector of nodes.}

\item{z}{An integer vector of nodes.}
}
\description{
Checks if the extraction procedure converged.
}
\keyword{internal}
