% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lrd}
\alias{lrd}
\title{Evaluates the exponential decay function at the current iteration.}
\usage{
lrd(epoch, learning_rate, decay_rate)
}
\arguments{
\item{epoch}{Integer for iteration.}

\item{learning_rate}{Double for the learning rate.}

\item{decay_rate}{Double for the decay rate.}
}
\description{
Evaluates the exponential decay function at the current iteration.
}
\keyword{internal}
