% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{main_search_decay_cpp}
\alias{main_search_decay_cpp}
\title{ECoHeN extraction procedure.}
\usage{
main_search_decay_cpp(
  alpha,
  B0,
  G_stats,
  learning_rate,
  decay_rate,
  adj_method,
  max_iter
)
}
\arguments{
\item{alpha}{Double of significance level.}

\item{B0}{Vector of vertices representing the candidate community.}

\item{G_stats}{List of graph statistics returned from \code{eval_G}.}

\item{learning_rate}{Double between 0 and 1 dictating the maximal allowance on the first iteration.}

\item{decay_rate}{Double between 0 and 1 dictating the maximal allowance after the first iteration.}

\item{adj_method}{String scalar for the desired multiple correction method. See \code{p.adjust.methods} for options.}

\item{max_iter}{Integer scalar for the maximum number of iterations before halting the extraction procedure. int max_iter, Rcpp::String adj_method}
}
\description{
ECoHeN extraction procedure.
}
