% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{political_blogs}
\alias{political_blogs}
\title{Political blogs network by Lada Adamic and Natalie Glance}
\format{
An igraph object with 1222 nodes, 16714 edges, and the following vertex attributes:
\describe{
\item{name}{name}
\item{source}{url}
\item{label}{numeric label}
\item{leaning}{political leaning attained via a text analysis of the blog content}
\item{color}{color associated with the political leaning}
}
}
\source{
\url{https://rdrr.io/github/barrpet/redwalk-r/man/polblogs.html}
}
\usage{
political_blogs
}
\description{
Political blogs network by Lada Adamic and Natalie Glance
}
\keyword{datasets}
