% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refine_ECoHeN.R
\name{refine_ECoHeN}
\alias{refine_ECoHeN}
\title{Refine extracted community structure.}
\usage{
refine_ECoHeN(
  G,
  ecs,
  beta = 0.5,
  min_size = 0,
  max_size = Inf,
  fully_connected = TRUE,
  trace = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{G}{igraph representation of graph.}

\item{ecs}{Extracted community structure. Resulting object from \code{cluster_ECoHeN} or a list of vectors.}

\item{beta}{The maximal overlap between clusters.}

\item{min_size}{The minimum cluster size to be considered.}

\item{max_size}{The maximum cluster size to be considered.}

\item{fully_connected}{Logical indicating whether or not to keep fully connected clusters.}

\item{trace}{Logical indicating whether or not to keep the trace.}

\item{verbose}{Logical indicating whether to print remaining communities.}
}
\value{
List containing the filtered community structure.
}
\description{
Refine extracted community structure.
}
\details{
Provided the extracted community structure, this function filters the
clusters to prioritize larger clusters such that no smaller cluster has a jaccard
similarity index larger than beta.
}
