% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_heterogeneous_sbm.R
\name{sample_heterogeneous_sbm}
\alias{sample_heterogeneous_sbm}
\title{Sample heterogeneous stochastic block model}
\usage{
sample_heterogeneous_sbm(
  C = 2,
  N = list(c(100, 50), c(100, 50)),
  B = c(350, 350),
  P = rbind(c(0.05, 0.05), c(0.05, 0.05)),
  R = rbind(c(0.25, 0.1), c(0.1, 0.2))
)
}
\arguments{
\item{C}{Number of communities.}

\item{N}{List of vectors describing number of nodes of each type in each community. There are as many elements as node types. Each element must be a vector of length C.}

\item{B}{Vector describing the number of background nodes of each type. They will be connected according to P.}

\item{P}{The between-community probability matrix. Must be a square, symmetric matrix with number of rows and columns equal to the number of node types. Provides with Bernoulli rate of connections between type-labeled nodes in different communities.}

\item{R}{The within-community additive probability matrix. Must be a square, symmetric matrix with number of rows and columns equal to the number of node types. Provides the additive increase to the Bernoulli rate of connections between type-labeled nodes in the same communities.}
}
\value{
An \code{igraph} graph.
}
\description{
Sampling from a heterogeneous stochastic block model of networks.
}
\details{
The function generates a random graph according to a heterogeneous stochastic block model. The default is a graph with two node types and three communities. The first community consists of 300 type I and 240 type II nodes. The second community consists of 200 type I and 160 type II nodes. The third community consists of 100 type I and 80 type II nodes. If a type I node is in the same community as a type II node, then their rate of connection is 0.30 (i.e. \code{P[1,2] + R[1,2]}); otherwise, they are connected with probability 0.05 (i.e. \code{P[1,2]}), for example.
}
\examples{
\dontrun{
sample_heterogeneous_sbm(
 C = 2,
 N = list(c(100, 50), c(100, 50)),
 B = c(350, 350),
 P = rbind(c(0.05, 0.05), c(0.05, 0.05)),
 R = rbind(c(0.25, 0.10), c(0.10, 0.20))
)
}
}
