% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{subset_matrix}
\alias{subset_matrix}
\title{Subset a matrix.}
\usage{
subset_matrix(M, rows, cols)
}
\arguments{
\item{M}{Integer matrix.}

\item{rows}{Integer vector of rows in C++ indices.}

\item{cols}{Integer vector of columns in C++ indices.}
}
\description{
Subset a matrix.
}
\keyword{internal}
