% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swap_if.R
\name{swap_if}
\alias{swap_if}
\title{Swap values in two columns based on the order of values in another two columns of a data frame.}
\usage{
swap_if(df, to_compare, to_swap)
}
\arguments{
\item{df}{Dataframe containing columns to compare and swap.}

\item{to_compare}{A vector of size two of column names. If the value of the first column is larger than the value of the second column, then the values in the columns \code{to_swap} are swapped.}

\item{to_swap}{A vector of size two of column names. If the value of the first column of \code{to_compare} is larger than the value of the second column of \code{to_compare}, then the values in the columns \code{to_swap} are swapped.}
}
\value{
Dataframe with the proper ordering of columns.
}
\description{
Swap values in two columns based on the order of values in another two columns of a data frame.
}
\examples{
df <- data.frame(a = c(1, 8, 3), b = c(1, 2, 27))
swap_if(df, to_compare = c('a', 'b'), to_swap = c('a', 'b'))
}
\keyword{internal}
