% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toy_example}
\alias{toy_example}
\title{Toy example of a simple, heterogeneous network}
\format{
An igraph object with 11 nodes, 15 edges, and the following vertex attributes:
\describe{
\item{name}{name}
\item{color}{color representing node type}
}
}
\source{
\url{https://rdrr.io/github/barrpet/redwalk-r/man/polblogs.html}
}
\usage{
toy_example
}
\description{
Toy example of a simple, heterogeneous network
}
\keyword{datasets}
