% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_equal.R
\name{vector_equal}
\alias{vector_equal}
\title{Check if two vectors are equivalent, avoiding the issue of recycling values.}
\usage{
vector_equal(x, y)
}
\arguments{
\item{x}{A vector.}

\item{y}{A vector.}
}
\value{
\code{TRUE} or \code{FALSE} depending on whether the vectors are equal.
}
\description{
Check if two vectors are equivalent, avoiding the issue of recycling values.
}
\examples{
x <- c(1, 2, 3)
y <- c(1, 2, 3, 1)
vector_equal(x, y)
all(x == y)
}
