% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vector_equals_string}
\alias{vector_equals_string}
\title{Construct logical vector by comparing a character vector to a string.}
\usage{
vector_equals_string(x, s)
}
\arguments{
\item{x}{Character vector.}

\item{s}{String.}
}
\description{
Construct logical vector by comparing a character vector to a string.
}
\keyword{internal}
