% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_heterogeneous_sbm.R
\name{viz_heterogeneous_sbm}
\alias{viz_heterogeneous_sbm}
\title{Visualize heterogeneous stochastic block model}
\usage{
viz_heterogeneous_sbm(G, block_communities = FALSE)
}
\arguments{
\item{G}{An igraph object attained from \code{sample_heterogeneous_sbm}}

\item{block_communities}{Should the adjacency be blocked by communities? Or node type?}
}
\value{
A \code{ggplot} object.
}
\description{
Provided a generated from the heterogeneous stochastic block model,
this function plots the adjacency matrix with the node type colored on the axes.
The plot includes nodes as tick marks where the axes are colored according to the node type.
}
\details{
See the documentation for \code{sample_heterogeneous_sbm} for more details.
}
