% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{which_true}
\alias{which_true}
\title{Indices of TRUE values in a logical vector.}
\usage{
which_true(x)
}
\arguments{
\item{x}{Logical vector.}
}
\description{
Indices of TRUE values in a logical vector.
}
\keyword{internal}
