% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_conductance.R
\name{compute_conductance}
\alias{compute_conductance}
\title{Compute conductance of a cut.}
\usage{
compute_conductance(G, vids)
}
\arguments{
\item{G}{igraph representation of graph.}

\item{vids}{Vector of vertex ids.}
}
\value{
A double between 0 and 1.
}
\description{
Compute conductance of a cut.
}
\keyword{internal}
